+++ pkg/handlers/helpers/authors/delete_author_on_moderation.go
@@ -3,22 +3,25 @@ package authors
 import (
 	"errors"
 
+	"github.com/Araks1255/mangacage/pkg/common/models"
 	"gorm.io/gorm"
 )
 
-func DeleteAuthorOnModeration(db *gorm.DB, authorOnModerationID, moderatorID uint) (code int, err error) {
-	result := db.Exec(
-		"DELETE FROM authors_on_moderation WHERE id = ? AND moderator_id = ?",
+func DeleteAuthorOnModeration(db *gorm.DB, authorOnModerationID, moderatorID uint) (deleted *models.AuthorOnModeration, code int, err error) {
+	var deletedAuthorOnModeration models.AuthorOnModeration
+
+	err = db.Raw(
+		"DELETE FROM authors_on_moderation WHERE id = ? AND moderator_id = ? RETURNING name, creator_id",
 		authorOnModerationID, moderatorID,
-	)
+	).Scan(&deletedAuthorOnModeration).Error
 
-	if result.Error != nil {
-		return 500, result.Error
+	if err != nil {
+		return nil, 500, err
 	}
 
-	if result.RowsAffected == 0 {
-		return 404, errors.New("автор на модерации не найден среди рассматриваемых вами")
+	if deletedAuthorOnModeration.Name == "" {
+		return nil, 404, errors.New("автор на модерации не найден среди рассматриваемых вами")
 	}
 
-	return 0, nil
+	return &deletedAuthorOnModeration, 0, nil
 }
